Delete bill API

+++ backend/src/api/controllers/bill.controller.ts
@@ -130,12 +130,12 @@ export const deleteBillController = async (req: Request, res: Response) => {
   const { id } = req.params;
 
   try {
-    // const { userId } = getAuth(req);
-    // if (!userId) {
-    //   let errorMessage = "Unauthorized: Invalid token";
-    //   console.log(errorMessage);
-    //   return res.status(401).json({ error: errorMessage });
-    // }
+    const { userId } = getAuth(req);
+    if (!userId) {
+      let errorMessage = "Unauthorized: Invalid token";
+      console.log(errorMessage);
+      return res.status(401).json({ error: errorMessage });
+    }
 
     const bill = await billService.deleteBillService(id as string);
     res.status(204).json({ bill });
